feat: route protection 1st part
